refactor: SDK 로딩 및 지도 초기화 분리
